MBS-11776 Extract claim to KubernetesReservationClaimer.kt
